refactor(enums): Centralize core data type enums
